Move router and middleware into messages folder
