+++ messages-api.js
@@ -3,7 +3,7 @@ const port = process.env.PORT || 3000
 
 app.use(
   require("body-parser").json(),
-  require("./messages-router"),
+  require("./messages/router"),
 )
 
 app.listen(port, () => console.log(`Listening on :${port}`))
